mapwidgethelper: add methods for editing MapLocations
updateCurrentDiveSiteCoordinates() should be called from
QML when a marker changed it's location. it emits the coordinatesChanged()
signal, which should be tracked in the MapWidget class. The MapWidget
class should emit the same signal to the MainWindow (Marble does that).

editMode is now a boolean property, which should put the QML map into
editing mode.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>